when server opens immediately force the domain to be localhost
-this prevents a situation where before you use your first cy.visit()
and you attempt to navigate to other http sites those sites will be
redirected into cypress
-due to the changes in the UI there is less chance of the user
modifying the URL directly which is what this primarily was beneficial
for
-and in fact we still catch those due to __cypress.unload
-the only thing “lost” is navigating directly to the URL after closing
the browser. in that case cypress will serve “/“ directly from your
local file system. but imo this is the least likely scenario and its
worth the trade off