fix: must wrap `npm` & `github` publish plugins individually
Incorrectly assumed it was possible to pass an array of plugins from a
plugin. User must explicitly pass wrapped `npm` and `github` publish
plugins. This is clunky and annoying. Will have to think about how the
`semantic-release` plugin system could be improved to allow for
composition like we’re trying to achieve here.