fix(withVersionHead): force the `gitHead` fallback algorithm
When the wrapped `plugin` is `@semantic-release/npm`, it will as
of `2.5.0` try to find a fallback `gitHead` itself. If it's found,
`gitHead` will be populated but incorrect for our purposes, yet
our fallback algorithm won't be used.

For now (at least while we remain locked to `semantic-release`
`~11.1.0`), force the fallback algorithm.