fixes #1940 force document.hasFocus to return what top does
this prevents situations where there are differences between running
within Cypress vs out of Cypress. normally the AUT is top and hasFocus
will return true, but in Cypress the AUT iframe has to literally be in
direct focus else it will return false.

hasFocus effectively “bubbles” when in iframes so children iframes in
focus cause their parents to receive the same value