Move external resources to peerDependencies
There's no need to set these as explicit dependencies as they're not used internally in a significant way, but to mirror the behavior of the user's libraries.  So in nearly ever case the user will already have a different version of these libraries as dependencies.  In the case that they do not, they will be prompted to install them.  This drastically reduces the dependencies for any project that is not upgrading their babel/webpack in line with what this library is hard set to.