Add ignoreTestFiles to configKeys
`ignoreTestFiles` is listed as a config option in the 
[guide](https://docs.cypress.io/guides/references/configuration.html#Global), 
but is not one of the `configKeys`.

This then causes Cypress to filter out `ignoreTestFiles` from the 
default config when filtering it to just the whitelisted values:
https://github.com/brianmcgue/cypress/blob/36760b8abbbf72692810f2dab266a13dd6fdbce7/packages/server/lib/project.coffee#L132

As such, when Cypress attempts to update the config with the plugin 
values `resolvedObj[key]` will return undefined here:
https://github.com/brianmcgue/cypress/blob/36760b8abbbf72692810f2dab266a13dd6fdbce7/packages/server/lib/config.coffee#L291
which will result in attempting to set values on `undefined`.

Alternatively, and possibly a good addition to this pull request as it 
stands currently, `setResolvedOn` should only recurse on non-array 
objects.
https://github.com/brianmcgue/cypress/blob/36760b8abbbf72692810f2dab266a13dd6fdbce7/packages/server/lib/config.coffee#L288
should become
```javascript
if _isObject(val) && !_.isArray(val)
```