feat: move default server configs to a separate package
The goal of this PR is to help extract the default server
config into a separate package that can be consumed and published
standalone. If it is required that we move this out of the repo,
I'm fine with that to. The goal of the PR is to show a potential
solution for https://github.com/cypress-io/cypress/issues/5217