use lodash instead of lodash/fp
The build was failing with webpack complaining that `lodash/fp` could not be found. Instead of spending any more time on it, I decided to just remove its use. I also worry that it adds more size to the bundle since it’s using a different branch of lodash so some amount of lodash is being duplicated in the bundle, though I couldn’t verify that because I could never get it building with `lodash/fp`