fix: use net-stubbing.d.ts in triple-slash reference
When referencing a .ts file (vs. .d.ts) file, an import of cypress in a
TypeScript project with a custom rootDir will lead to this error

```
error TS6059: File
'<snip>/foo/node_modules/cypress/types/net-stubbing.ts' is not under
'rootDir' '<snip>foo/bar'. 'rootDir' is expected to contain all source
files.
```

Using a .d.ts file is in line with the other references in index.d.ts
and fixes this compiler error.