Update npm-react tests to run via CLI and use Xvfb
Previously, these tests were failing because, in CI, Xvfb is required to run Chrome, and these tests do not run via the CLI. Normally the CLI starts Xvfb.

Updated to route the tests via the CLI via a `--run-project` option on `open-ct`. I'm assuming there will be a `run-ct` command eventually? But this works for now.