init: fixed startup delay issue
This change fixes an issue that caused the startup to be delayed by 500ms.
We used `app.on('ready')` without checking `app.isReady()`, however the
'ready' event may already have fired at this point and thus we never
resolved the Promise waiting for it.
Using `app.whenReady` fixes this and now that Promise resolves before
the delay expires.

Related links:

https://www.electronjs.org/docs/api/app#event-ready
https://www.electronjs.org/docs/api/app#appwhenready