Update dependencies with security vulnerabilities
Removed "konfig": "0.2.1" packages as this package was last published 6 years ago and there is no way to update its dependency 'js-yaml' as the repo for this package does not exists
Upgraded 'http-server' package to the latest version which uses the latest version of 'http-proxy'. Previous version of 'http-proxy' had security vulnaribilities.

More details:
https://github.com/advisories/GHSA-6x33-pw7p-hmpq
https://github.com/advisories/GHSA-8j8c-7jfh-h6hx