inject timers code into aut instead of being a part of the driver
We now wrap timers in the AUT because if we do it in the
driver/top frame, any uncaught errors thrown in the timer callbacks
get picked up by the top frame's 'error' handler instead of the AUT's.
We need to wrap the timer callbacks in the AUT itself for errors to
propagate properly.