feat: Support running cypress as node process (#16505)
This change checks if the environment variable
ELECTRON_RUN_AS_NODE is set. If set to true, the cypress
process is started as a normal Node.js process rather than an
electron process.

Consequently, the process does not require a display server to
run so starting a display server is skipped.