feat: add support for using Playwright browsers
Introduces basic support for using the browsers bundles with the
Microsoft Playwright package as target browser for running Cypress tests

Introduces a new `browser`-type `playwright` which depends on
the `name`-property of the `FoundBrowser`-type to determine which of the
three support Playwright browsers should be opened by the Cypress runner
the support browsers are: `firefox`, `chromium`, and `webkit`.

A new automation client has been implemented for Playwright supports is
not feature complete, most of the tests of the Cypress example tests
should work as expected (e.g. navigate, local storage) but some features
are not tested or working, which are:

  - Download handling (I would need a simple example project for this)
  - Video recording

Potentially, features which don't match the same behaviour as the default
browser support of Cypres are:

  - Screenshots; its creating screenshots but might not generate the same
    images as the other browsers

refs #16996, #6422