fix: plugin file in angular schematics (#17141)
Runnin g`ng add @cypress/schematic --defaults --skip-confirmation` in a new Angular 12 application results in a broken setup for cypress,
as the plugin file does not have any export.

When running `ng e2e`, Cypress complains with:

```
The plugins file is missing or invalid.

Your `pluginsFile` is set to `/Users/ced-pro/Code/angular/angular-cli-diff/ponyracer/cypress/plugins/index.js`, but either the file is missing, it contains a syntax error, or threw an error when required. The `pluginsFile` must be a `.js`, `.ts`, or `.coffee` file.

Or you might have renamed the extension of your `pluginsFile`. If that's the case, restart the test runner.

Please fix this, or set `pluginsFile` to `false` if a plugins file is not necessary for your project.
```

This commit reverts the changes to the plugin file made in 96a9db4204f50a7605a5e51d51584d27aa9df164
and fixes the issue.