test: fix server unit files test
This was a bit tricky as this is clearly an integration test, just
looking at the SUT (`lib/gui/files.ts`) makes that obvious as there
isn't much going on in there directly.

Additionally during the setup we did create a project which was entirely
unnecessary for these tests. That step required stubbing of the
ProjectBase `getConfig` which isn't a direct dep of the module we are
testing (a _smell_ when we consider that we are _unit_ testing).

Instead I stubbed `openProject.getConfig` directly and thus the tests
pass and are more isolate WRT to the part of the system they are
touching/testing.