feat: make grepFilterSpecs work with negative grep
Since find-test-names 1.18, it's possible to output the
full test names, i.e. prepended with the parent
suite names.
This allows grepFilterSpecs to filter
spec files when negative grep matches test but
not suite and vice versa.

Benefit:
The log is less cluttered, especially
with a lot of specs.
The speed improvement is minor,
about 1.5 seconds per filtered spec file.