misc: Where possible allow Firefox and Chrome to atomically allocate ports
Firefox and Chrome let the system allocate a port when started with
 --remote-debugging-port=0
and announce the port on stderr afterwards where it is now parsed from.

Do something similar for Firefox's Marionette port and parse it
from a message that gets written to stdout.

This doesn't seem possible with Firefox's
`devtools.debugger.remote-port` so at least allow it to be overridden
via preferences so users can implement their own race free port allocation
mechanism when / where required:

	setupNodeEvents: (on, config) => {
            on('before:browser:launch', (browser, opts) => {
                opts.preferences['devtools.debugger.remote-port'] = 12345
                return opts;
            })
      	}

Co-authored-by: Emily Rohrbough <emilyrohrbough@users.noreply.github.com>